+++ punchpipe/control/scheduler.py
@@ -1,4 +1,5 @@
 import inspect
+import itertools
 from datetime import datetime
 
 from prefect import get_run_logger
@@ -11,8 +12,42 @@ def generic_scheduler_flow_logic(
     query_ready_files_func, construct_child_file_info, construct_child_flow_info, pipeline_config_path,
         update_input_file_state=True, new_input_file_state="progressed",
         session=None, reference_time: datetime | None = None,
-        args_dictionary: dict = {}
-):
+        args_dictionary: dict = {},
+        children_are_one_to_one: bool = False,
+    ):
+    """
+    Implement the core logic of each scheduler flow.
+
+    Parameters
+    ----------
+    query_ready_files_func
+        A function that returns a list of lists, where each of the inner lists is a group of files that will be the
+        inputs to one flow
+    query_ready_files_func
+        A function that generates the child File entries for one group/flow
+    construct_child_flow_info
+        A function that generates the Flow entry for one group
+    pipeline_config_path
+        The config path
+    update_input_file_state
+        Whether to change the state of the input files
+    new_input_file_state
+        The new state to assign to each input file
+    session
+        A database Session
+    reference_time
+        Some time of observation time associated with this scheduling run. The meaning is defined by flow-specific
+        functions passed into this function.
+    args_dictionary: dict
+         Values in this dictionary are passed directly to the 
+         `query_ready_files_func`, `query_ready_files_func`, and `construct_child_flow_info` functions
+    children_are_one_to_one
+        By default, for each group of input files, it is assumed that all inputs together produce all the output
+        files, and FileRelationships are generated accordingly. In a case where a batch of input files are to be
+        processed in one flow, this assumption doesn't hold. When this flag is set to True, it is assumed each input
+        file connects to only one output file (at the corresponding position in the list of child File objects).
+    """
+
     logger = get_run_logger()
     pipeline_config = load_pipeline_configuration(pipeline_config_path)
 
@@ -70,7 +105,10 @@ def generic_scheduler_flow_logic(
             session.commit()
 
             # create a file relationship between the prior and next levels
-            for parent_file in parent_files:
-                for child_file in children_files:
-                    session.add(FileRelationship(parent=parent_file.file_id, child=child_file.file_id))
+            if children_are_one_to_one:
+                iterable = zip(parent_files, children_files)
+            else:
+                iterable = itertools.product(parent_files, children_files)
+            for parent_file, child_file in iterable:
+                session.add(FileRelationship(parent=parent_file.file_id, child=child_file.file_id))
             session.commit()